@@ -2,22 +2,34 @@ package com.emmanuelhmar.booklistingapp;
 
 import android.os.AsyncTask;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import android.util.Log;
 import android.widget.ListView;
 
+import java.util.ArrayList;
 import java.util.List;
 
+import androidx.appcompat.app.AppCompatActivity;
+import butterknife.BindView;
+import butterknife.ButterKnife;
+
 public class MainActivity extends AppCompatActivity {
+    private final static String LOG_TAG = MainActivity.class.getName();
     private final String JSON_URL = "https://www.googleapis.com/books/v1/volumes?q=android&maxResults=1";
+    private static BookAdapter bookAdapter;
+    @BindView(R.id.list_item) ListView listView;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+gi
+        ButterKnife.bind(this);
 
-        ListView listView = findViewById(R.id.list_item);
+        bookAdapter = new BookAdapter(this, new ArrayList<Book>());
 
+        listView.setAdapter(bookAdapter);
 
+        new BookAsyncTask().execute(JSON_URL);
 
     }
 
@@ -26,6 +38,18 @@ public class MainActivity extends AppCompatActivity {
         protected List<Book> doInBackground(String... strings) {
 
             return QueryUtils.fetchBookData(strings[0]);
+        }
+
+        @Override
+        protected void onPostExecute(List<Book> books) {
+            bookAdapter.clear();
+
+            Log.i(LOG_TAG, "Books: " + books);
+
+            if (books!= null && !books.isEmpty()) {
+                bookAdapter.addAll(books);
+            }
+
         }
     }
 
